requirements.txt
bs4
requests
pillow
pytesseract
cv
numpy
matplotlib
tabulate
scikit-learn